Separate TrySplitNamespaceAndType tests into class-specific file
