+++ src/Razor/Microsoft.AspNetCore.Razor.Language/test/DefaultRazorTagHelperBinderPhaseTest.cs
@@ -1342,26 +1342,6 @@ namespace Microsoft.AspNetCore.Razor.Language
             Assert.True(result);
         }
 
-        [Theory]
-        [InlineData("", false, "", "")]
-        [InlineData(".", true, "", "")]
-        [InlineData("Foo", true, "", "Foo")]
-        [InlineData("SomeProject.Foo", true, "SomeProject", "Foo")]
-        [InlineData("SomeProject.Foo<Bar>", true, "SomeProject", "Foo<Bar>")]
-        [InlineData("SomeProject.Foo<Bar.Baz>", true, "SomeProject", "Foo<Bar.Baz>")]
-        [InlineData("SomeProject.Foo<Bar.Baz>>", true, "", "SomeProject.Foo<Bar.Baz>>")]
-        [InlineData("SomeProject..Foo<Bar>", true, "SomeProject.", "Foo<Bar>")]
-        public void TrySplitNamespaceAndType_WorksAsExpected(string fullTypeName, bool expectedResult, string expectedNamespace, string expectedTypeName)
-        {
-            // Arrange & Act
-            var result = RazorCSharpUtilities.TrySplitNamespaceAndType(fullTypeName, out var @namespace, out var typeName);
-
-            // Assert
-            Assert.Equal(expectedResult, result);
-            Assert.Equal(expectedNamespace, DefaultRazorTagHelperBinderPhase.ComponentDirectiveVisitor.GetTextSpanContent(@namespace, fullTypeName));
-            Assert.Equal(expectedTypeName, DefaultRazorTagHelperBinderPhase.ComponentDirectiveVisitor.GetTextSpanContent(typeName, fullTypeName));
-        }
-
         private static RazorSourceDocument CreateComponentTestSourceDocument(string content, string filePath = null)
         {
             var sourceDocument = TestRazorSourceDocument.Create(content, filePath: filePath);
